Made Website Responsive , Added Final Functionalities

+++ init/index.js
@@ -15,7 +15,7 @@ async function main() {
 
 
 const initDB=async()=>{
-    await Listing.deleteMany({});
+    await Listing.deleteMany({ owner: "6a4dbfdd38bdb6b6d3fe25a8"});
     AllData.data=AllData.data.map((obj)=>({...obj,owner:"6a4dbfdd38bdb6b6d3fe25a8"}));
    await  Listing.insertMany(AllData.data);
 }
